+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Added
+
+- Nodepool nodes are labeled with nodepool id on AWS using `giantswarm.io/machine-pool`.
+
 ## [1.40.0] - 2021-09-24
 
 ### Added
 
-- Nodepool nodes are labeled with nodepool id on Azure.
+- Nodepool nodes are labeled with nodepool id on Azure using `giantswarm.io/machine-pool`.
 
 ### Changed